Add shared project
