+++ ARAppSourceCode/PlaneARViewer/PlaneARViewer.sln
@@ -7,7 +7,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlaneARViewer", "PlaneARVie
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "..\Test\Test.csproj", "{F33CA401-97FE-45AB-8A9F-4A2E34F1B6AE}"
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "SharedAirplaneFinder", "..\SharedAirplaneFinder\SharedAirplaneFinder.shproj", "{7DFEEBAB-5365-4A4F-A8A5-0E7337F83728}"
+EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		..\SharedAirplaneFinder\SharedAirplaneFinder.projitems*{7dfeebab-5365-4a4f-a8a5-0e7337f83728}*SharedItemsImports = 13
+		..\SharedAirplaneFinder\SharedAirplaneFinder.projitems*{be099bee-d4e6-4e79-8164-b1cc80cf31cc}*SharedItemsImports = 4
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|iPhone = Debug|iPhone
