+++ ARAppSourceCode/SharedAirplaneFinder/SharedAirplaneFinder.projitems
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>7dfeebab-5365-4a4f-a8a5-0e7337f83728</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>SharedAirplaneFinder</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)AirplaneFinder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DataManager.cs" />
+  </ItemGroup>
+</Project>
